Fix very annoying scrolling when editing bug
Cells are reused.  When destroyed to be recycled, the data value input by user
was not being retained if scrolled out of view.  This would crash the app as data values
were not present when the model was used again and index section & row values for
the recycled cell would be nil.

Implement fix by guarding row and section values are not nil and also implementing TableView
delegate method 'didEndDisplayingCell'.  In this method, I take the textField input value by
the user and update the model with it.  When the cell is recycled and then re-displayed, the
input value is retained.

Also add border to texField and re-implement edit style 'none'