Persist data straight after seeding the table view with placeholder data
I wasn't doing this previously, which meant a user action like 'edit' had to be performed
before the data was persisted.  If no data was persisted then the pie chart would not work etc as
there was nothing to work with and the placeholder data would seed next time round.  Now no action
needs to be performed after seeding.